 \docType{data}
    \name{camcap}
    \alias{camcap}
    \title{The Cambridge Cohort from Ross-Adams et al. (2015)}
    \format{a \code{ExpressionSet} instance, 1 row per Illumina Huamnv4 probe.}
    \source{
      GSE70768
    }
    \description{
      Object imported using GEOquery and curated to remove variables introduced by GEO. See Vignette for details.
    }
    \author{
Mark Dunning
    }
    \value{ExpressionSet object for the dataset}
    \usage{camcap}
    \keyword{datasets}